% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_physiology_data.R
\name{physiology_data}
\alias{physiology_data}
\title{physiology_data}
\usage{
physiology_data()
}
\value{
A list with metadata and data
}
\description{
physiology_data
}
\examples{
\dontrun{
physiology_data()
}
}
